stanley-proxy/requirements.txt
aiohttp>=3.9
asyncpg>=0.29
python-dotenv>=1.0
